Model class

+++ src/main/java/org/example/model/Input.java
@@ -2,9 +2,11 @@ package org.example.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Input {
     private String fromAccount;
     private String toAccount;
-    private Double sumRemittance;
+    private BigDecimal sumRemittance;
 }
